taskmgr/views.py
from django.shortcuts import render
from rest_framework import generics
from rest_framework import viewsets
from rest_framework import permissions
from .models import Task
from .serializers import TaskSerializer

# Create your views here.

# ////////////////////////////////////////////////////////// 1


class TaskList(generics.ListCreateAPIView):
    queryset = Task.objects.all()
    serializer_class = TaskSerializer


class TaskDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Task.objects.all()
    serializer_class = TaskSerializer

# ////////////////////////////////////////////////////////// 2
